New docker image for Go+Docker+Heroku
We want the Docker tools, to talk to a remote Docker server, for use
with Circle CI's `setup_remote_docker` so that we can build Docker
images when running in Docker.  We also want Go, but now can just use
upstream Golang images, and we add Heroku as before, and our packages.

This image drops the user to a non-root runtime user, with no way back
to root.  Not a problem: if we need more privs, we can rebuild the
container.