One tiny hack and can access local Docker from this image
It's not wonderful, but it should work for our common workflows.
For CI usage, we use whatever the CI system wants, to talk to the Docker
which it's willing to make available to us (for Circle CI, that's
Docker-in-Docker, `dind`).  When on your own laptop, you might instead
just want to be able to access "the docker I'm running in".

To do this, bind-mount the Docker Unix-domain socket into the image.  Of
course, the runtime user needs to be able to access that socket.

Those mac users who are using docker-machine to launch Docker will
almost certainly be using boot2docker, which creates
`/var/run/docker.sock` mode 0660 ownership `root:docker`, where the gid
of `docker` is 100.

Inside stretch, gid 100 is `users`, which is an eminently sensible gid
for our runtime user to belong to.

With this change:

    docker run -v /var/run/docker.sock:/var/run/docker.sock domainr/ci:tag

is sufficient for `docker ps` to work inside the built image.

While I'm here, add `less` to the installed packages before I go crazy.